firmware: qcom_scm: Wire up System Performance Dynamic Monitoring calls
The System Performance Dynamic Monitoring (SPDM) system is a "method
and system for managing bandwidth demand for a variable bandwidth
processing element in a portable computing device" (quote from US
patent application publication 2015/0161070 A1).

This is used by various SoCs to control CPU and/or GPU bandwidth
scaling and limiting in hardware and is made for another driver to
use this API in order to configure the SPDM block, which has to be
done through secure world communication.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@somainline.org>