interconnect: qcom: Add System Performance Dynamic Monitoring driver
The System Performance Dynamic Monitoring (SPDM) controls the bandwidth
of the Cache Coherent Interconnect (CCI) on some Qualcomm SoCs, like
MSM8956, MSM8996, MSM8998, SDM630/660 for which support is implemented
and others, which I've had no way to test so the configuration has not
been added.

Implement an interconnect driver that talks to the SPDM through secure
world calls and permits to achieve CCI scaling on CPU frequency change.

Signed-off-by: *NOT YET*